+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/annotators/BaseAnnotator.kt
@@ -11,7 +11,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.impl.source.tree.LeafPsiElement
 
 
-abstract class BaseAnnotator : Annotator, DumbAware {
+abstract class BaseAnnotator : Annotator {
 
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
 
@@ -23,7 +23,6 @@ abstract class BaseAnnotator : Annotator, DumbAware {
             holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
                 .range(range)
                 .textAttributes(kind)
-                .needsUpdateOnTyping()
                 .create()
         }